+++ sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = 'budgeter-v34';
+const CACHE_NAME = 'budgeter-v36';
 
 const FILES_TO_CACHE = [
   './',